refactor(front): Rename objectItem to ObjectCard
