+++ ui/components/imageWorkspace/src/components/ImageInspector/ObjectCard.svelte
